requirements.txt
click==8.1.3
Flask==2.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
Pillow==9.2.0
python-docx==0.8.11
Werkzeug==2.2.0
zipp==3.8.1
